kafka/confluent/helper/requirements.txt
pyyaml>=5.4.1
Jinja2>=2.11.3
